pyproject.toml
[tool.poetry]
name = "rest-auth-multitoken"
version = "0.1.0"
description = "Add multiple tokens support to django-rest-auth."
authors = ["makerGeek <moetez.93@gmail.com>"]
readme = "readme.md"
homepage = "https://github.com/makerGeek/rest-auth-multitoken"
repository = "https://github.com/makerGeek/rest-auth-multitoken"
license = "MIT"


[tool.poetry.dependencies]
python = "^3.9"
django-rest-auth = "^0.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
